+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.
 
+## 0.4.0 (2026-06-02)
+
+
+### Features
+
+* Add publisher and workload ID validation
+* Add support for new claims .md location
+
+
+### Documentation
+
+* Add .md files describing claims supported by PES
+
 ## 0.3.0 (2026-05-29)